add cats feature

+++ routes/web.php
@@ -1,4 +1,5 @@
 <?php
+use App\Http\Middleware\CheckAdmin;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,11 +21,11 @@ Auth::routes();
 Route::middleware(['auth'])->group(function () {
     Route::get('/home', 'ArticleController@index')->name('home');
     Route::group(['prefix' => 'article'], function () {
-        Route::get('/create', 'ArticleController@add')->name('create-article');
-        Route::post('/create', 'ArticleController@store')->name('create-action-article');
-        Route::get('{id}/edit', 'ArticleController@update')->name('update-article');
-        Route::post('{id}/update', 'ArticleController@updateaction')->name('update-action-article');
-        Route::delete('{id}/delete', 'ArticleController@delete')->name('delete-action-article');
+        Route::get('/create', 'ArticleController@add')->name('create-article')->middleware(CheckAdmin::class);
+        Route::post('/create', 'ArticleController@store')->name('create-action-article')->middleware(CheckAdmin::class);
+        Route::get('{id}/edit', 'ArticleController@update')->name('update-article')->middleware(CheckAdmin::class);
+        Route::post('{id}/update', 'ArticleController@updateaction')->name('update-action-article')->middleware(CheckAdmin::class);
+        Route::delete('{id}/delete', 'ArticleController@delete')->name('delete-action-article')->middleware(CheckAdmin::class);
         Route::get('{id}/save', 'ArticleController@save')->name('save-article');
         Route::get('{id}/unsave', 'ArticleController@unsave')->name('unsave-article');
         Route::get('{id}/publish', 'ArticleController@publish')->name('publish-article');
@@ -33,11 +34,11 @@ Route::middleware(['auth'])->group(function () {
         Route::get('{id}', 'ArticleController@showone')->name('view-article');
     });
     Route::group(['prefix' => 'cat'], function () {
-        Route::get('/create', 'CatController@add')->name('create-cat');
-        Route::post('/create', 'CatController@store')->name('create-action-cat');
-        Route::delete('{id}/delete', 'CatController@delete')->name('delete-action-cat');
+        Route::get('/create', 'CatController@add')->name('create-cat')->middleware(CheckAdmin::class);
+        Route::post('/create', 'CatController@store')->name('create-action-cat')->middleware(CheckAdmin::class);
+        Route::delete('{id}/delete', 'CatController@delete')->name('delete-action-cat')->middleware(CheckAdmin::class);
         Route::get('{id}/articles', 'CatController@getarticles')->name('get-articles-by-cat');
     });
-    Route::get('my-articles','UserController@showMyArticles')->name('view-my-article');
+    Route::get('my-articles','UserController@showMyArticles')->name('view-my-article')->middleware(CheckAdmin::class);
     Route::get('my-saved-articles','UserController@showMySavedArticles')->name('view-my-saved-article');
 });
